+++ cli/golang_internal_testsuite/testsuite_impl/basic_datastore_test/basic_datastore_test_.go
@@ -6,17 +6,20 @@
 package basic_datastore_test
 
 import (
+	"context"
 	"fmt"
-	"github.com/kurtosis-tech/example-microservice/datastore/datastore_service_client"
+	"github.com/golang/protobuf/ptypes/empty"
+	"github.com/kurtosis-tech/example-datastore-server/api/golang/datastore_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/services"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
+	"google.golang.org/grpc"
+	"time"
 )
 
 const (
-	datastoreImage                        = "kurtosistech/example-microservices_datastore"
 	datastoreServiceId services.ServiceID = "datastore"
 	datastorePort                         = 1323
 	testKey                               = "test-key"
@@ -38,27 +41,37 @@ func (test BasicDatastoreTest) Configure(builder *testsuite.TestConfigurationBui
 	builder.WithSetupTimeoutSeconds(60).WithRunTimeoutSeconds(60)
 }
 
-func (test BasicDatastoreTest) Setup(networkCtx *networks.NetworkContext) (networks.Network, error) {
+func (test BasicDatastoreTest) Setup(networkCtx *networks.NetworkContext) (network networks.Network, returnErr error) {
+	ctx := context.Background()
 
-	datastoreContainerConfigSupplier := getDatastoreContainerConfigSupplier()
+	datastoreContainerConfigSupplier := test.getDatastoreContainerConfigSupplier()
 
 	serviceContext, hostPortBindings, err := networkCtx.AddService(datastoreServiceId, datastoreContainerConfigSupplier)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred adding the datastore service")
 	}
 
-	datastoreClient := datastore_service_client.NewDatastoreClient(serviceContext.GetIPAddress(), datastorePort)
+	datastoreClient, datastoreClientConnCloseFunc, err := newDatastoreClient(serviceContext.GetIPAddress())
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a new datastore client for service with ID '%v' and IP address '%v'", datastoreServiceId, serviceContext.GetIPAddress())
+	}
+	defer func() {
+		err = datastoreClientConnCloseFunc()
+		returnErr = stacktrace.Propagate(err, "An error occurred closing GRPC client")
+	}()
 
-	err = datastoreClient.WaitForHealthy(waitForStartupMaxPolls, waitForStartupDelayMilliseconds)
+	err = waitDatastoreServiceForHealthy(ctx, datastoreClient, waitForStartupMaxPolls, waitForStartupDelayMilliseconds)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred waiting for the datastore service to become available")
 	}
 
 	logrus.Infof("Added datastore service with host port bindings: %+v", hostPortBindings)
-	return networkCtx, nil
+	return networkCtx, returnErr
 }
 
-func (test BasicDatastoreTest) Run(network networks.Network) error {
+func (test BasicDatastoreTest) Run(network networks.Network) (returnErr error)  {
+	ctx := context.Background()
+
 	// Necessary because Go doesn't have generics
 	castedNetwork := network.(*networks.NetworkContext)
 
@@ -67,47 +80,105 @@ func (test BasicDatastoreTest) Run(network networks.Network) error {
 		return stacktrace.Propagate(err, "An error occurred getting the datastore service info")
 	}
 
-	datastoreClient := datastore_service_client.NewDatastoreClient(serviceContext.GetIPAddress(), datastorePort)
+	datastoreClient, datastoreClientConnCloseFunc, err := newDatastoreClient(serviceContext.GetIPAddress())
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred creating a new datastore client for service with ID '%v' and IP address '%v'", datastoreServiceId, serviceContext.GetIPAddress())
+	}
+	defer func() {
+		err = datastoreClientConnCloseFunc()
+		returnErr = stacktrace.Propagate(err, "An error occurred closing GRPC client")
+	}()
 
 	logrus.Infof("Verifying that key '%v' doesn't already exist...", testKey)
-	exists, err := datastoreClient.Exists(testKey)
+	existsArgs := &datastore_rpc_api_bindings.ExistsArgs{
+		Key: testKey,
+	}
+	existsResponse, err := datastoreClient.Exists(ctx, existsArgs)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred checking if the test key exists")
 	}
-	if exists {
+	if existsResponse.GetExists() {
 		return stacktrace.NewError("Test key should not exist yet")
 	}
 	logrus.Infof("Confirmed that key '%v' doesn't already exist", testKey)
 
 	logrus.Infof("Inserting value '%v' at key '%v'...", testKey, testValue)
-	if err := datastoreClient.Upsert(testKey, testValue); err != nil {
+	upsertArgs := &datastore_rpc_api_bindings.UpsertArgs{
+		Key: testKey,
+		Value: testValue,
+	}
+	if _, err = datastoreClient.Upsert(ctx, upsertArgs); err != nil {
 		return stacktrace.Propagate(err, "An error occurred upserting the test key")
 	}
 	logrus.Infof("Inserted value successfully")
 
 	logrus.Infof("Getting the key we just inserted to verify the value...")
-	value, err := datastoreClient.Get(testKey)
+	getArgs := &datastore_rpc_api_bindings.GetArgs{
+		Key: testKey,
+	}
+	getResponse, err := datastoreClient.Get(ctx, getArgs)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the test key after upload")
 	}
-	if value != testValue {
-		return stacktrace.NewError("Returned value '%v' != test value '%v'", value, testValue)
+	if getResponse.GetValue() != testValue {
+		return stacktrace.NewError("Returned value '%v' != test value '%v'", getResponse.GetValue(), testValue)
 	}
 	logrus.Info("Value verified")
-	return nil
+	return returnErr
 }
 
 // ====================================================================================================
 //                                       Private helper functions
 // ====================================================================================================
-func getDatastoreContainerConfigSupplier() func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
+func (test BasicDatastoreTest) getDatastoreContainerConfigSupplier() func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
 	containerConfigSupplier  := func(ipAddr string, sharedDirectory *services.SharedPath) (*services.ContainerConfig, error) {
 		containerConfig := services.NewContainerConfigBuilder(
-			datastoreImage,
+			test.datastoreImage,
 		).WithUsedPorts(
 			map[string]bool{fmt.Sprintf("%v/tcp", datastorePort): true},
 		).Build()
 		return containerConfig, nil
 	}
 	return containerConfigSupplier
+}
+
+func newDatastoreClient(datastoreIp string) (datastore_rpc_api_bindings.DatastoreServiceClient, func() error, error) {
+	datastoreURL := fmt.Sprintf(
+		"%v:%v",
+		datastoreIp,
+		datastorePort,
+	)
+
+	conn, err := grpc.Dial(datastoreURL, grpc.WithInsecure())
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred dialling the datastore container via its URL")
+	}
+
+	datastoreServiceClient := datastore_rpc_api_bindings.NewDatastoreServiceClient(conn)
+
+	return datastoreServiceClient, conn.Close, nil
+}
+
+func waitDatastoreServiceForHealthy(ctx context.Context, client datastore_rpc_api_bindings.DatastoreServiceClient, retries uint32, retriesDelayMilliseconds uint32) error {
+
+	var (
+		emptyArgs = &empty.Empty{}
+		err error
+	)
+
+	for i := uint32(0); i < retries; i++ {
+		_, err = client.IsAvailable(ctx, emptyArgs)
+		if err == nil  {
+			break
+		}
+		time.Sleep(time.Duration(retriesDelayMilliseconds) * time.Millisecond)
+	}
+
+	if err != nil {
+		return stacktrace.Propagate(err,
+			"The datastore service didn't return a success code, even after %v retries with %v milliseconds in between retries",
+			 retries, retriesDelayMilliseconds)
+	}
+
+	return nil
 }